Adding new flag: finished (no more pulls needed)
Split 'next' flag into two, 'next' and 'finished'. The former now meaning that that is the value one will need to set to the '?next=' URL argument in order to retrieve the following messages in the sequence, whereas the latter is there merely to inform clients that no more pulls are necessary currently as all messages so far have been pulled. This way, whenever the client decides to check with the server that new messages have arrived, they can provide the 'next' value and get exactly what they want.